encapsulate different quiz views into components
